+++ frontend/src/components/MenuAppBar.jsx
@@ -1,34 +1,15 @@
-import AccountCircle from '@mui/icons-material/AccountCircle';
+import { ProfileMenu, Sidebar } from '#components';
+import HomeIcon from '@mui/icons-material/Home';
+import InventoryIcon from '@mui/icons-material/Inventory';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
-import IconButton from '@mui/material/IconButton';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
 import Toolbar from '@mui/material/Toolbar';
-import HomeIcon from '@mui/icons-material/Home';
-import InventoryIcon from '@mui/icons-material/Inventory';
 import Typography from '@mui/material/Typography';
-import { Sidebar } from '#components';
-import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 export default function MenuAppBar({ user, logoutAsync, isLoading }) {
-    const [anchorEl, setAnchorEl] = useState(null);
     const navigate = useNavigate();
-
-    const handleLogoutClick = async () => {
-        await logoutAsync();
-        handleClose();
-    };
-
-    const handleMenu = event => {
-        setAnchorEl(event.currentTarget);
-    };
-
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
 
     const sidebarMenu = [
         { text: 'Home', link: '/home', role: 'CUSTOMER', icon: <HomeIcon /> },
@@ -38,6 +19,12 @@ export default function MenuAppBar({ user, logoutAsync, isLoading }) {
             role: 'STAFF',
             icon: <InventoryIcon />
         }
+    ];
+
+    const profileMenu = [
+        { text: 'Profile', link: '/profile' },
+        { text: 'My Account', link: '/account' },
+        { text: 'Logout', onClick: logoutAsync }
     ];
 
     return (
@@ -52,48 +39,7 @@ export default function MenuAppBar({ user, logoutAsync, isLoading }) {
                     >
                         IoT Bay
                     </Typography>
-                    {user && (
-                        <div>
-                            <IconButton
-                                size='large'
-                                aria-label='account of current user'
-                                aria-controls='menu-appbar'
-                                aria-haspopup='true'
-                                onClick={handleMenu}
-                                color='inherit'
-                            >
-                                <AccountCircle />
-                            </IconButton>
-                            <Menu
-                                id='menu-appbar'
-                                anchorEl={anchorEl}
-                                anchorOrigin={{
-                                    vertical: 'top',
-                                    horizontal: 'right'
-                                }}
-                                keepMounted
-                                transformOrigin={{
-                                    vertical: 'top',
-                                    horizontal: 'right'
-                                }}
-                                open={Boolean(anchorEl)}
-                                onClose={handleClose}
-                            >
-                                <MenuItem onClick={handleClose}>
-                                    Profile
-                                </MenuItem>
-                                <MenuItem onClick={handleClose}>
-                                    My account
-                                </MenuItem>
-                                <MenuItem
-                                    disabled={isLoading}
-                                    onClick={handleLogoutClick}
-                                >
-                                    Logout
-                                </MenuItem>
-                            </Menu>
-                        </div>
-                    )}
+                    {user && <ProfileMenu items={profileMenu} />}
                     {!user && (
                         <Button
                             color='inherit'
